Fixed zero hits issue

+++ PixelDecoder/PixelDecoder.h
@@ -24,7 +24,7 @@ public:
     virtual ~Decoder() { }
     int open(std::string file, int chanMulti);
     void decodeRoc32(uint32_t line, int chanID, int count);
-    void decodeRoc64(uint32_t line, int chanID, int count);
+    void decodeRoc64(uint64_t line, int chanID, int count);
     void graph(std::string file);
 };
